feat: Added some fixes to getBlock()
In MetalRenegades, the height of the gooey is 1. Most of the time, the
NavGraphSystem would not be able to find a suitable WalkableBlock
(because it would be looking underground).
Since entities have a span, the exact block on which they are standing
cannot be ascertained directly by just checking the block. Even the
surrounding blocks need to be checked.